refactor: moved configurations into their own file
